centralize place to put reset error
