+++ src/actions/common.js
@@ -1,0 +1,10 @@
+export const RESET_FAILED = "RESET_FAILED";
+export function resetFailedReason(){
+	return (dispatch) => {
+		dispatch({
+			type:RESET_FAILED,
+			reason:"",
+			receivedAt: Date.now(),
+		});
+	};
+}
