+++ src/actions/dataset.js
@@ -144,14 +144,3 @@ export function datasetDeleted(datasetId){
 			});
 	};
 }
-
-export const RESET_FAILED = "RESET_FAILED";
-export function resetFailedReason(){
-	return (dispatch) => {
-		dispatch({
-			type:RESET_FAILED,
-			reason:"",
-			receivedAt: Date.now(),
-		});
-	};
-}
